LEARNIX-71 DB Modell angepasst Services und Endpunkte am neuen Modell angepasst

+++ src/main/java/de/szut/learnixback/services/QuizQuestionService.java
@@ -4,8 +4,8 @@ import de.szut.learnixback.entities.QuizQuestion;
 import de.szut.learnixback.repositories.QuizQuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class QuizQuestionService {
@@ -13,29 +13,33 @@ public class QuizQuestionService {
     @Autowired
     private QuizQuestionRepository quizQuestionRepository;
 
-    public QuizQuestion createQuizQuestion(QuizQuestion quizQuestion) {
-        return quizQuestionRepository.save(quizQuestion);
-    }
-
     public List<QuizQuestion> getAllQuizQuestions() {
         return quizQuestionRepository.findAll();
     }
 
-    public QuizQuestion getQuizQuestionById(int id) {
-        return quizQuestionRepository.findById(id).orElse(null);
+    public Optional<QuizQuestion> getQuizQuestionById(Long id) {
+        return quizQuestionRepository.findById(id);
     }
 
-    public QuizQuestion updateQuizQuestion(int id, QuizQuestion updatedQuizQuestion) {
-        QuizQuestion existingQuizQuestion = quizQuestionRepository.findById(id).orElse(null);
-        if (existingQuizQuestion != null) {
-            existingQuizQuestion.setQuiz(updatedQuizQuestion.getQuiz());
-            // Weitere Aktualisierungen hier, falls erforderlich
-            return quizQuestionRepository.save(existingQuizQuestion);
+    public QuizQuestion createQuizQuestion(QuizQuestion quizQuestion) {
+        return quizQuestionRepository.save(quizQuestion);
+    }
+
+    public QuizQuestion updateQuizQuestion(Long id, QuizQuestion updatedQuizQuestion) {
+        if (quizQuestionRepository.existsById(id)) {
+            updatedQuizQuestion.setQuizQuestionId(id);
+            return quizQuestionRepository.save(updatedQuizQuestion);
+        } else {
+            return null;
         }
-        return null;
     }
 
-    public void deleteQuizQuestion(int id) {
-        quizQuestionRepository.deleteById(id);
+    public boolean deleteQuizQuestion(Long id) {
+        if (quizQuestionRepository.existsById(id)) {
+            quizQuestionRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
