+++ src/main/java/de/szut/learnixback/services/QuizService.java
@@ -4,8 +4,8 @@ import de.szut.learnixback.entities.Quiz;
 import de.szut.learnixback.repositories.QuizRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class QuizService {
@@ -13,28 +13,33 @@ public class QuizService {
     @Autowired
     private QuizRepository quizRepository;
 
-    public Quiz createQuiz(Quiz quiz) {
-        return quizRepository.save(quiz);
-    }
-
     public List<Quiz> getAllQuizzes() {
         return quizRepository.findAll();
     }
 
-    public Quiz getQuizById(int id) {
-        return quizRepository.findById(id).orElse(null);
+    public Optional<Quiz> getQuizById(Long id) {
+        return quizRepository.findById(id);
     }
 
-    public Quiz updateQuiz(int id, Quiz updatedQuiz) {
-        Quiz existingQuiz = quizRepository.findById(id).orElse(null);
-        if (existingQuiz != null) {
-            existingQuiz.setLections(updatedQuiz.getLections()); // Falls nötig, um andere Eigenschaften zu aktualisieren, füge weitere Aktualisierungen hinzu
-            return quizRepository.save(existingQuiz);
+    public Quiz createQuiz(Quiz quiz) {
+        return quizRepository.save(quiz);
+    }
+
+    public Quiz updateQuiz(Long id, Quiz updatedQuiz) {
+        if (quizRepository.existsById(id)) {
+            updatedQuiz.setQuizId(id);
+            return quizRepository.save(updatedQuiz);
+        } else {
+            return null;
         }
-        return null;
     }
 
-    public void deleteQuiz(int id) {
-        quizRepository.deleteById(id);
+    public boolean deleteQuiz(Long id) {
+        if (quizRepository.existsById(id)) {
+            quizRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
